requirements.txt
PyYAML==6.0
requests==2.28.1
aiohttp==3.8.3
packaging==21.3
